move IIPrange to common
