+++ src/queue/authentication/IQueueSASSignatureValues.ts
@@ -3,7 +3,7 @@ import {
   truncatedISO8061Date
 } from "../../common/utils/utils";
 import { SASProtocol } from "../../common/authentication/IAccountSASSignatureValues";
-import { IIPRange, ipRangeToString } from "./IIPRange";
+import { IIPRange, ipRangeToString } from "../../common/authentication/IIPRange";
 
 /**
  * ONLY AVAILABLE IN NODE.JS RUNTIME.
