+++ src/table/authentication/ITableSASSignatureValues.ts
@@ -3,7 +3,7 @@ import {
   truncatedISO8061Date
 } from "../../common/utils/utils";
 import { SASProtocol } from "../../common/authentication/IAccountSASSignatureValues";
-import { IIPRange, ipRangeToString } from "./IIPRange";
+import { IIPRange, ipRangeToString } from "../../common/authentication/IIPRange";
 
 /**
  * ITableSASSignatureValues is used to help generating Table service SAS tokens for containers or tables.
